+++ module/data/item/shield.js
@@ -13,7 +13,7 @@ export class ShieldItemData extends GeneralItemData {
       ...super.defineSchema('shield'),
       bonus: new fields.NumberField({ initial: 1, integer: true }),
       equipped: new fields.BooleanField({ initial: false }),
-      minStrength: new fields.StringField({ initial: '' }),
+      minStrength: new fields.NumberField({ initial: 0, integer: true }),
       notes: new fields.StringField({ initial: '' }),
     };
   }