+++ module/dramaticScene/torgeternityCombat.js
@@ -84,6 +84,9 @@ export default class TorgCombat extends Combat {
         const hand = combatant.actor.getDefaultHand();
         if (hand) hand.setFlag('torgeternity', 'disablePlayCards', true)
       }
+
+      // Check if the turn marker should be immediately displayed
+      combatant.token?.object?.renderFlags.set({ refreshTurnMarker: true })
     }
   }
 
@@ -377,7 +380,7 @@ export default class TorgCombat extends Combat {
     console.log('Drama Stymied', faction)
     this.#sendDramaChat('stymied', faction);
     for (const actor of this.getFactionActors(faction))
-      actor.applyStymiedState('stymied', faction);
+      actor.increaseStymied(faction);
   }
 
   async dramaSurge(faction) {