+++ module/test-dialog.js
@@ -181,7 +181,8 @@ export class TestDialog extends HandlebarsApplicationMixin(ApplicationV2) {
       CONFIG.torgeternity.concentrationSkills.includes(context.test.skillName)) {
       context.test.concentratingModifier = myActor.statusModifiers.concentrating;
     }
-    context.test.requiresConcentration = this.test.itemId && myActor.items.get(this.test.itemId)?.requiresConcentration;
+    const testItem = this.test.itemId && myActor.items.get(this.test.itemId);
+    context.test.requiresConcentration = testItem?.requiresConcentration;
 
     // Set Modifiers for Vehicles
     if (this.test.testType === 'chase') {
@@ -209,7 +210,7 @@ export class TestDialog extends HandlebarsApplicationMixin(ApplicationV2) {
     const targets = this.options.useTargets ? Array.from(game.user.targets) : [];
     context.test.targetPresent = !!targets.length;
     const MULTITARGET = [0, 0, -2, -4, -6, -8, -10];
-    context.test.targetsModifier ||= MULTITARGET[targets.length] ?? 0;
+    context.test.targetsModifier ||= MULTITARGET[testItem?.hasBlastTrait ? 1 : targets.length] ?? 0;
 
     if (context.test.targetPresent && context.test.testType !== 'soak') {
       context.test.targetAll = targets.map(token => oneTestTarget(token, this.test.applySize));
@@ -328,7 +329,6 @@ export function dummyTestTargets() {
   return [{
     dummyTarget: true,
     amountBD: 0,
-    addBDs: 0,
     bdDamageSum: 0,
     damage: 0,
   }];
@@ -367,7 +367,6 @@ export function oneTestTarget(token, applySize) {
       armor: actor.defenses.armor,
       armorTraits: [],
       amountBD: 0,
-      addBDs: 0,
       bdDamageSum: 0,
       // then vehicle specifics
       defenses: {
@@ -404,7 +403,6 @@ export function oneTestTarget(token, applySize) {
       darknessModifier: actor.statusModifiers.darkness,
       isConcentrating: actor.isConcentrating,
       amountBD: 0,
-      addBDs: 0,
       bdDamageSum: 0,
       defenses: {
         ...damageDefenses,
@@ -467,8 +465,8 @@ export function TestDialogLabel(test) {
       result = `${test.skillName} ${game.i18n.localize('torgeternity.chatText.test')}  `;
   }
   if (test.itemId) {
-    const itemName = fromUuidSync(test.actor, { strict: false })?.items.get(test.itemId).name;
-    if (itemName) result += ` (${itemName})`;
+    const item = fromUuidSync(test.actor, { strict: false })?.items.get(test.itemId);
+    if (item) result += `<br>(${item.name}${item.system?.traits?.has('trademark') ? '\u2122' : ''})`;
   }
   return result;
 }